+++ ft_memcpy.c
@@ -6,15 +6,17 @@
 /*   By: mbazirea <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/31 15:10:26 by mbazirea          #+#    #+#             */
-/*   Updated: 2022/11/04 15:34:41 by mbazirea         ###   ########.fr       */
+/*   Updated: 2022/11/05 17:07:21 by mbazirea         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memcpy(void *dst, const void *src, int size)
+#include "libft.h"
+
+void	*ft_memcpy(void *dst, const void *src, size_t size)
 {
 	char	*src2;
 	char	*dst2;
-	int		i;
+	size_t	i;
 
 	src2 = (char *) src;
 	dst2 = (char *) dst;
